Updated AudioControls
Changed the way recorded audio can be gathered from AudioControls. 

Modified the way the audio display code works. It now only displays the last second or so of recorded audio. NON-URGENT: There is a value that defines "silence" for the display scheme.  It needs to be tweaked to visually "hide" some amount of background noise.